+++ DockDoor/consts.swift
@@ -23,7 +23,7 @@ extension Defaults.Keys {
     static let screenCaptureCacheLifespan = Key<CGFloat>("screenCaptureCacheLifespan", default: 60 )
     static let uniformCardRadius = Key<Bool>("uniformCardRadius", default: true )
     static let tapEquivalentInterval = Key<CGFloat>("tapEquivalentInterval", default: 1.5 )
-    static let previewHoverAction = Key<PreviewHoverAction>("previewHoverAction", default: PreviewHoverAction.none )
+    static let previewHoverAction = Key<PreviewHoverAction>("previewHoverAction", default: .none )
     
     static let showAnimations = Key<Bool>("showAnimations", default: true )
     static let enableWindowSwitcher = Key<Bool>("enableWindowSwitcher", default: true )
@@ -35,26 +35,31 @@ extension Defaults.Keys {
     static let Int64maskAlternate = Key<Int>("Int64maskAlternate", default: 524576 )
     static let UserKeybind = Key<UserKeyBind>("UserKeybind", default: UserKeyBind(keyCode: 48, modifierFlags: Defaults[.Int64maskControl]))
     
+    static let showAppName = Key<Bool>("showAppName", default: true)
+    static let appNameStyle = Key<AppNameStyle>("appNameStyle", default: .default )
+    
     static let showWindowTitle = Key<Bool>("showWindowTitle", default: true )
-    static let windowTitleDisplayCondition = Key<WindowTitleDisplayCondition>("windowTitleDisplayCondition", default: .always )
-    static let windowTitlePosition = Key<WindowTitlePosition>("windowTitlePosition", default: WindowTitlePosition.bottomLeft )
-    static let windowTitleStyle = Key<WindowTitleStyle>("windowTitleStyle", default: .default )
-    static let trafficLightButtonsVisibility = Key<TrafficLightButtonsVisibility>("trafficLightButtonsVisibility", default: .dimmedOnWindowHover )
+    static let windowTitleDisplayCondition = Key<WindowTitleDisplayCondition>("windowTitleDisplayCondition", default: .all)
+    static let windowTitleVisibility = Key<WindowTitleVisibility>("windowTitleVisibility", default: .whenHoveringPreview)
+    static let windowTitlePosition = Key<WindowTitlePosition>("windowTitlePosition", default: .bottomLeft )
+    
+    static let trafficLightButtonsVisibility = Key<TrafficLightButtonsVisibility>("trafficLightButtonsVisibility", default: .dimmedOnPreviewHover )
+    static let trafficLightButtonsPosition = Key<TrafficLightButtonsPosition>("trafficLightButtonsPosition", default: .topLeft)
 }
 
 enum WindowTitleDisplayCondition: String, CaseIterable, Defaults.Serializable {
-    case always = "always"
+    case all = "all"
     case dockPreviewsOnly = "dockPreviewsOnly"
     case windowSwitcherOnly = "windowSwitcherOnly"
     
     var localizedName: String {
         switch self {
-        case .always:
-            String(localized: "Always", comment: "Preview window title condition option")
+        case .all:
+            String(localized: "Dock Previews & Window Switcher", comment: "Preview window title display condition option")
         case .dockPreviewsOnly:
-            String(localized: "When Showing Dock Tile Previews", comment: "Preview window title condition option")
+            String(localized: "Dock Previews only", comment: "Preview window title condition display option")
         case .windowSwitcherOnly:
-            String(localized: "When Using Window Switcher", comment: "Preview window title condition option")
+            String(localized: "Window Switcher only", comment: "Preview window title condition display option")
         }
     }
 }
@@ -63,6 +68,7 @@ enum WindowTitlePosition: String, CaseIterable, Defaults.Serializable {
     case bottomLeft
     case bottomRight
     case topRight
+    case topLeft
     
     var localizedName: String {
         switch self {
@@ -72,20 +78,19 @@ enum WindowTitlePosition: String, CaseIterable, Defaults.Serializable {
             String(localized: "Bottom Right", comment: "Preview window title position option")
         case .topRight:
             String(localized: "Top Right", comment: "Preview window title position option")
+        case .topLeft:
+            String(localized: "Top Left", comment: "Preview window title position option")
         }
     }
 }
 
-enum WindowTitleStyle: String, CaseIterable, Defaults.Serializable {
-    case hidden
+enum AppNameStyle: String, CaseIterable, Defaults.Serializable {
     case `default`
     case embedded
     case popover
     
     var localizedName: String {
         switch self {
-        case .hidden:
-            String(localized: "Hidden", comment: "Preview title style option")
         case .default:
             String(localized: "Default", comment: "Preview title style option")
         case .embedded:
@@ -96,22 +101,56 @@ enum WindowTitleStyle: String, CaseIterable, Defaults.Serializable {
     }
 }
 
+enum WindowTitleVisibility: String, CaseIterable, Defaults.Serializable {
+    case whenHoveringPreview
+    case alwaysVisible
+    
+    var localizedName: String {
+        switch self {
+        case .whenHoveringPreview:
+            String(localized: "When hovering over the preview", comment: "Window title visibility option")
+        case .alwaysVisible:
+            String(localized: "Always visible", comment: "Window title visibility option")
+        }
+    }
+}
+
 enum TrafficLightButtonsVisibility: String, CaseIterable, Defaults.Serializable {
     case never
-    case dimmedOnWindowHover
-    case fullOpacityOnWindowHover
+    case dimmedOnPreviewHover
+    case fullOpacityOnPreviewHover
     case alwaysVisible
     
     var localizedName: String {
         switch self {
         case .never:
             String(localized: "Never visible", comment: "Traffic light buttons visibility option")
-        case .dimmedOnWindowHover:
+        case .dimmedOnPreviewHover:
             String(localized: "On window hover; Dimmed until button hover", comment: "Traffic light buttons visibility option")
-        case .fullOpacityOnWindowHover:
+        case .fullOpacityOnPreviewHover:
             String(localized: "On window hover; Full opacity", comment: "Traffic light buttons visibility option")
         case .alwaysVisible:
             String(localized: "Always visible; Full opacity", comment: "Traffic light buttons visibility option")
+        }
+    }
+}
+
+enum TrafficLightButtonsPosition: String, CaseIterable, Defaults.Serializable {
+    case topLeft
+    case topRight
+    case bottomRight
+    case bottomLeft
+    
+    var localizedName: String {
+        switch self {
+        case .topLeft:
+            String(localized: "Top Left", comment: "Traffic light buttons position option")
+        case .topRight:
+            String(localized: "Top Right", comment: "Traffic light buttons position option")
+        case .bottomRight:
+            String(localized: "Bottom Right", comment: "Traffic light buttons position option")
+        case .bottomLeft:
+            String(localized: "Bottom Left", comment: "Traffic light buttons position option")
         }
     }
 }
@@ -124,11 +163,11 @@ enum PreviewHoverAction: String, CaseIterable, Defaults.Serializable {
     var localizedName: String {
         switch self {
         case .none:
-            String(localized: "No hover action", comment: "Window popup hover action option")
+            String(localized: "No action", comment: "Window popup hover action option")
         case .tap:
-            String(localized: "Simulate a click", comment: "Window popup hover action option")
+            String(localized: "Simulate a click (open the window)", comment: "Window popup hover action option")
         case .previewFullSize:
-            String(localized: "See a preview of the window", comment: "Window popup hover action option")
+            String(localized: "See a large preview of the window", comment: "Window popup hover action option")
         }
     }
 }